requirements.txt
python-dateutil~=2.8.1
PyYAML~=5.4.1
requests~=2.25.1
Deprecated~=1.2.11
pytest~=6.2.2
responses~=0.13
python-dotenv~=0.19.0
setuptools_scm~=6.2
invoke~=1.6.0
pylint=2.11.1